refactor(go.d/snmp-traps): internalize profile metric runtime
