+++ src/go/plugin/go.d/collector/snmp_traps/internal/catalog/profile.go
@@ -227,8 +227,7 @@ type Epoch struct {
 	validationStockProfile string
 }
 
-// NewEpoch returns an empty epoch for callers that assemble static profile
-// definitions before the runtime-ownership migration is complete.
+// NewEpoch returns an empty epoch for callers that assemble static trap definitions.
 func NewEpoch() *Epoch {
 	return &Epoch{
 		trapsByOID:        make(map[string]*TrapDef),
@@ -241,13 +240,6 @@ func NewEpoch() *Epoch {
 
 // AddTraps validates and adds static trap definitions to an unpublished epoch.
 func (idx *Epoch) AddTraps(traps []*TrapDef) error { return idx.addTraps(traps) }
-
-// AddMetricDefinitions validates and adds manually assembled static metric
-// definitions to an unpublished epoch. Definitions accumulated by earlier
-// calls are one logical bundle for this temporary compatibility API.
-func (idx *Epoch) AddMetricDefinitions(rules []MetricRule, charts []MetricChart) error {
-	return idx.addProfileMetrics(rules, charts, true)
-}
 
 // PrepareTrap compiles one manually assembled trap definition.
 func PrepareTrap(td *TrapDef) error {
@@ -399,8 +391,6 @@ type MetricDefinitions struct {
 }
 
 // Definitions returns the selected static metric definitions and their charts.
-// A nil ruleNames slice returns the complete snapshot for the temporary manual
-// assembly bridge; runtime callers pass their explicit selection.
 func (idx *Epoch) Definitions(ruleNames []string) (MetricDefinitions, error) {
 	if idx == nil {
 		return MetricDefinitions{}, errors.New("profile index not available")
@@ -410,29 +400,20 @@ func (idx *Epoch) Definitions(ruleNames []string) (MetricDefinitions, error) {
 	}
 	idx.mu.RLock()
 	defer idx.mu.RUnlock()
-	if ruleNames != nil {
-		defs := MetricDefinitions{
-			RulesByName: make(map[string]*MetricRule, len(ruleNames)),
-			ChartsByID:  make(map[string]*MetricChart, len(ruleNames)),
-		}
-		for _, name := range ruleNames {
-			rule := idx.metricRulesByName[name]
-			if rule == nil {
-				continue
-			}
-			defs.RulesByName[name] = rule
-			if chart := idx.metricChartsByID[rule.Output.Chart]; chart != nil {
-				defs.ChartsByID[chart.ID] = chart
-			}
-		}
-		return defs, nil
-	}
 	defs := MetricDefinitions{
-		RulesByName: make(map[string]*MetricRule, len(idx.metricRulesByName)),
-		ChartsByID:  make(map[string]*MetricChart, len(idx.metricChartsByID)),
+		RulesByName: make(map[string]*MetricRule, len(ruleNames)),
+		ChartsByID:  make(map[string]*MetricChart, len(ruleNames)),
 	}
-	maps.Copy(defs.RulesByName, idx.metricRulesByName)
-	maps.Copy(defs.ChartsByID, idx.metricChartsByID)
+	for _, name := range ruleNames {
+		rule := idx.metricRulesByName[name]
+		if rule == nil {
+			continue
+		}
+		defs.RulesByName[name] = rule
+		if chart := idx.metricChartsByID[rule.Output.Chart]; chart != nil {
+			defs.ChartsByID[chart.ID] = chart
+		}
+	}
 	return defs, nil
 }
