+++ src/go/plugin/go.d/collector/snmp_traps/internal/profilemetrics/charts_test.go
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package profilemetrics
+
+import (
+	"testing"
+)
+
+func TestProfileMetricDiagnosticChartUsesTemplateLocalContext(t *testing.T) {
+	chart := profileMetricDiagnosticChart()
+	if chart.Context != "profile_metric_diagnostics" {
+		t.Fatalf("diagnostic chart context = %q, want profile_metric_diagnostics", chart.Context)
+	}
+}
